+++ src/tools/redpepper/tools/entrypoint.py
@@ -145,6 +145,7 @@ def install_step_keypair_agent(
         REDPEPPER_CONFIG_DIR, "agent.d", "01-step-ca-certificate.yaml"
     ),
     steppath: str = DEFAULT_STEP_PATH,
+    stepbinary: str = os.path.join(REDPEPPER_INSTALL_DIR, ".local", "bin", "step"),
     check_hostname: bool = False,
     install_renew_cron_job: bool = sys.platform == "linux",
     renew_schedule: str = "0 */2 * * *",
@@ -156,7 +157,13 @@ def install_step_keypair_agent(
     from .step_ca_keypair import create_step_ca_keypair, install_cert_renew_cron_job
 
     create_step_ca_keypair(
-        steppath, "RedPepper Agent", cert_file, key_file, ca_url, root_fingerprint
+        steppath,
+        stepbinary,
+        "RedPepper Agent",
+        cert_file,
+        key_file,
+        ca_url,
+        root_fingerprint,
     )
     with open(config_file, "w") as f:
         f.write(f'tls_cert_file: "{cert_file}"\n')
@@ -168,6 +175,7 @@ def install_step_keypair_agent(
     if install_renew_cron_job:
         install_cert_renew_cron_job(
             steppath,
+            stepbinary,
             cert_file,
             key_file,
             "redpepper-renew-agent-cert",
@@ -186,6 +194,7 @@ def install_step_keypair_manager(
         REDPEPPER_CONFIG_DIR, "manager.d", "01-step-ca-certificate.yaml"
     ),
     steppath: str = DEFAULT_STEP_PATH,
+    stepbinary: str = os.path.join(REDPEPPER_INSTALL_DIR, ".local", "bin", "step"),
     check_hostname: bool = False,
     install_renew_cron_job: bool = sys.platform == "linux",
     renew_schedule: str = "0 */2 * * *",
@@ -202,10 +211,16 @@ def install_step_keypair_manager(
     )
 
     if root_fingerprint is None:
-        root_fingerprint = get_step_ca_root_fingerprint(steppath)
+        root_fingerprint = get_step_ca_root_fingerprint(steppath, stepbinary)
 
     create_step_ca_keypair(
-        steppath, "RedPepper Manager", cert_file, key_file, ca_url, root_fingerprint
+        steppath,
+        stepbinary,
+        "RedPepper Manager",
+        cert_file,
+        key_file,
+        ca_url,
+        root_fingerprint,
     )
     with open(config_file, "w") as f:
         f.write(f'tls_cert_file: "{cert_file}"\n')
@@ -217,6 +232,7 @@ def install_step_keypair_manager(
     if install_renew_cron_job:
         install_cert_renew_cron_job(
             steppath,
+            stepbinary,
             cert_file,
             key_file,
             "redpepper-renew-manager-cert",
@@ -235,6 +251,7 @@ def install_step_keypair_manager_api(
         REDPEPPER_CONFIG_DIR, "manager.d", "01-step-ca-api-certificate.yaml"
     ),
     steppath: str = DEFAULT_STEP_PATH,
+    stepbinary: str = os.path.join(REDPEPPER_INSTALL_DIR, ".local", "bin", "step"),
     install_renew_cron_job: bool = sys.platform == "linux",
     renew_schedule: str = "0 */2 * * *",
     renew_post_cmd: str = "systemctl restart redpepper-manager",
@@ -250,10 +267,16 @@ def install_step_keypair_manager_api(
     )
 
     if root_fingerprint is None:
-        root_fingerprint = get_step_ca_root_fingerprint(steppath)
+        root_fingerprint = get_step_ca_root_fingerprint(steppath, stepbinary)
 
     create_step_ca_keypair(
-        steppath, "RedPepper Manager API", cert_file, key_file, ca_url, root_fingerprint
+        steppath,
+        stepbinary,
+        "RedPepper Manager API",
+        cert_file,
+        key_file,
+        ca_url,
+        root_fingerprint,
     )
     with open(config_file, "w") as f:
         f.write(f'api_tls_cert_file: "{cert_file}"\n')
@@ -263,6 +286,7 @@ def install_step_keypair_manager_api(
     if install_renew_cron_job:
         install_cert_renew_cron_job(
             steppath,
+            stepbinary,
             cert_file,
             key_file,
             "redpepper-renew-manager-api-cert",